+++ infra/italynorth/_modules/backoffice_app_service/main.tf
@@ -77,6 +77,15 @@ module "itn_sign_backoffice_app_roles" {
       roles = {
         secrets = "reader"
       }
+    },
+    {
+      name                = data.azurerm_key_vault.sign_kv.name
+      resource_group_name = data.azurerm_key_vault.sign_kv.resource_group_name
+      description         = "Allow ${module.itn_sign_backoffice_app.app_service.app_service.name} to write secrets to ${data.azurerm_key_vault.sign_kv.name}"
+      has_rbac_support    = true
+      roles = {
+        secrets = "writer"
+      }
     }
   ]
 }
@@ -95,6 +104,15 @@ module "itn_sign_backoffice_app_staging_roles" {
       has_rbac_support    = true
       roles = {
         secrets = "reader"
+      }
+    },
+    {
+      name                = data.azurerm_key_vault.sign_kv.name
+      resource_group_name = data.azurerm_key_vault.sign_kv.resource_group_name
+      description         = "Allow ${module.itn_sign_backoffice_app.app_service.app_service.slot.name} to write secrets to ${data.azurerm_key_vault.sign_kv.name}"
+      has_rbac_support    = true
+      roles = {
+        secrets = "writer"
       }
     }
   ]